// frontend/src/lib/api/types.gen.ts
export type CreateExpositorSchema = {
    nome: string;
    descricao?: string | null;
};